refactor(desktop): extract the link-preview runtime owner out of main.ts
main.ts is the composition seam, not the runtime: the title cache, the
one-hop pinned curl transport, the hidden title-window queue, the durable
preview cache, and the IPC surface all move to link-preview-runtime.ts
(review blocking item 3). The module takes app/BrowserWindow/session as a
deps object so it stays unit-testable without Electron, and registers its
two handles (hermes:fetchLinkTitle, hermes:fetchLinkPreview) through
registerLinkPreviewRuntime — main.ts constructs and registers, nothing more.
No behavior change: the handles resolve identically, the tier-2 window keeps
its guarded session, and the preview flow is unchanged apart from the
thumbnail leg landing in the previous commit.
