+++ apps/desktop/electron/main.ts
@@ -229,21 +229,7 @@ import { snapHudBounds } from './hud-snap'
 import { createHudSnapShortcut } from './hud-snap-shortcut'
 import { buildHudWindowUrl } from './hud-url'
 import { resolveHudWindowing } from './hud-windowing'
-import {
-  fetchWithGuardedRedirects,
-  HostRateLimiter,
-  type HttpHopResponse,
-  type LinkPreviewIo,
-  type LinkPreviewResult,
-  LinkPreviewStore,
-  PREVIEW_HOST_SPACING_MS,
-  PREVIEW_MAX_CONCURRENT,
-  PREVIEW_MAX_REDIRECTS,
-  PREVIEW_TTL_MS,
-  type PreviewPersistence,
-  resolveLinkPreview
-} from './link-preview'
-import { createLinkTitleWindow, guardLinkTitleSession, readLinkTitleWindowTitle } from './link-title-window'
+import { createLinkPreviewRuntime, registerLinkPreviewRuntime, TITLE_BYTE_BUDGET, TITLE_USER_AGENT } from './link-preview-runtime'
 import { ensureMainWindow } from './main-window-lifecycle'
 import {
   assertManagedUpdatePreflightClear,
@@ -5337,387 +5323,12 @@ function filenameFromUrl(rawUrl, fallback = 'image') {
   }
 }
 
-// Link title resolution — curl (tier 1) → hidden BrowserWindow (tier 2).
-const titleCache = new Map()
-const titleInflight = new Map()
-const TITLE_CACHE_LIMIT = 500
-const TITLE_BYTE_BUDGET = 96 * 1024
-const TITLE_TIMEOUT_MS = 5000
-
-// Browser-shaped UA — many bot-walled sites (GetYourGuide, Cloudflare-protected
-// pages) refuse anything that doesn't look like a real Chrome.
-const TITLE_USER_AGENT =
-  'Mozilla/5.0 (Macintosh; Intel Mac OS X 14_6_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/136.0.0.0 Safari/537.36'
-
-const TITLE_ERROR_RE =
-  /\b(access denied|attention required|captcha|error|forbidden|just a moment|request blocked|too many requests)\b/i
-
-const HTML_ENTITIES = { amp: '&', lt: '<', gt: '>', quot: '"', apos: "'", nbsp: ' ', '#39': "'" }
-
-// Tier-2 renderer fallback config. Only invoked when curl came back empty or
-// matched TITLE_ERROR_RE — keeps cold/CDN-cached pages on the cheap path.
-const RENDER_TITLE_MAX_CONCURRENT = 2
-const RENDER_TITLE_TIMEOUT_MS = 8000
-const RENDER_TITLE_GRACE_MS = 700
-
-let linkTitleSession = null
+// Link title + link preview resolution live in link-preview-runtime.ts (the
+// runtime owner: pinned one-hop curl transport, hidden title window, durable
+// preview cache, and the IPC surface). main.ts composes it and keeps only the
+// pieces other subsystems share.
 let oauthSession = null
-let renderTitleInFlight = 0
-const renderTitleQueue = []
 
-function canonicalTitleCacheKey(rawUrl) {
-  const value = String(rawUrl || '').trim()
-
-  if (!value) {
-    return ''
-  }
-
-  try {
-    const url = new URL(value)
-    const host = url.hostname.replace(/^www\./i, '').toLowerCase()
-    const pathname = url.pathname === '/' ? '/' : url.pathname.replace(/\/+$/, '') || '/'
-
-    return `${host}${pathname}${url.search || ''}`
-  } catch {
-    return value
-  }
-}
-
-function cacheTitle(key, title) {
-  if (titleCache.size >= TITLE_CACHE_LIMIT) {
-    titleCache.delete(titleCache.keys().next().value)
-  }
-
-  titleCache.set(key, title)
-}
-
-function decodeHtmlEntities(value) {
-  return value
-    .replace(/&(amp|lt|gt|quot|apos|nbsp|#39);/gi, (_, k) => HTML_ENTITIES[k.toLowerCase()] ?? '')
-    .replace(/&#x([0-9a-f]+);/gi, (_, hex) => String.fromCodePoint(parseInt(hex, 16) || 32))
-    .replace(/&#(\d+);/g, (_, dec) => String.fromCodePoint(parseInt(dec, 10) || 32))
-}
-
-function parseHtmlTitle(html) {
-  const raw = html.match(/<title[^>]*>([\s\S]*?)<\/title>/i)?.[1]
-
-  return raw ? decodeHtmlEntities(raw).replace(/\s+/g, ' ').trim() : ''
-}
-
-// ─── Tier-1 page fetch (one resolver for the title fetcher AND link previews) ─
-// curl is driven ONE HOP at a time and every hop is re-validated by
-// fetchWithGuardedRedirects (link-preview.ts). The old `--location` followed
-// any 30x to an arbitrary destination with no guard, which turned a vetted
-// public link into a fetch of loopback/RFC1918/169.254.x.x. `--include` puts
-// the status line and headers on stdout so the hop loop can read them; the
-// byte budget covers the header block too, which keeps the cap honest.
-// TITLE_TIMEOUT_MS is the budget for the WHOLE walk — the same wall clock the
-// single `--max-time` governed before.
-
-function splitHopResponse(raw: string): HttpHopResponse {
-  const separator = raw.indexOf('\r\n\r\n')
-
-  if (separator < 0) {
-    // No header block: malformed response, treat as a transport failure.
-    return { status: 0, location: '', body: '' }
-  }
-
-  const head = raw.slice(0, separator)
-  const statusLine = head.split('\n')[0] ?? ''
-  const status = Number(statusLine.match(/\b(\d{3})\b/)?.[1] ?? 0)
-  const location = head.match(/^location:[ \t]*(.*)$/im)?.[1]?.trim() ?? ''
-
-  return { status, location, body: raw.slice(separator + 4) }
-}
-
-function curlOneHop(url: string, timeoutMs: number, pinnedAddresses: string[] = []): Promise<HttpHopResponse> {
-  return new Promise(resolve => {
-    // DNS pinning (shape from upstream #63171's fetchPinnedLinkTitle): the
-    // connection is bound to the addresses the guard just vetted via
-    // --resolve, so curl never resolves the hostname itself and an attacker
-    // controlling DNS cannot swap the answer between the verdict and the
-    // request. Every vetted address is pinned; curl fails over among them
-    // without a fresh lookup. When the hop is a literal-IP URL there is no
-    // name to pin and the list is empty.
-    const port = new URL(url).port || (new URL(url).protocol === 'https:' ? '443' : '80')
-    const hostname = new URL(url).hostname.replace(/^\[|\]$/g, '')
-    const resolveArgs = pinnedAddresses.flatMap(address => ['--resolve', `${hostname}:${port}:${address}`])
-
-    const args = [
-      '--silent',
-      '--show-error',
-      '--include',
-      '--max-time',
-      String(Math.max(1, Math.ceil(timeoutMs / 1000))),
-      '--connect-timeout',
-      '4',
-      ...resolveArgs,
-      '--user-agent',
-      TITLE_USER_AGENT,
-      '--header',
-      'Accept: text/html,application/xhtml+xml;q=0.9,*/*;q=0.5',
-      '--header',
-      'Accept-Language: en-US,en;q=0.7',
-      '--header',
-      'Accept-Encoding: identity',
-      '--raw',
-      url
-    ]
-
-    const child = spawn('curl', args, hiddenWindowsChildOptions({ stdio: ['ignore', 'pipe', 'ignore'] }))
-    const chunks: Buffer[] = []
-    let bytes = 0
-
-    child.stdout.on('data', chunk => {
-      if (bytes >= TITLE_BYTE_BUDGET) {
-        return
-      }
-
-      const buffer = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk)
-      const remaining = TITLE_BYTE_BUDGET - bytes
-      const next = buffer.length > remaining ? buffer.subarray(0, remaining) : buffer
-      chunks.push(next)
-      bytes += next.length
-    })
-
-    child.on('error', () => resolve({ status: 0, location: '', body: '' }))
-    child.on('close', () => resolve(splitHopResponse(Buffer.concat(chunks).toString('utf8'))))
-  })
-}
-
-function resolveHostAddresses(hostname: string): Promise<string[]> {
-  return new Promise(resolve => {
-    try {
-      dns.lookup(hostname, { all: true }, (error, addresses) => {
-        resolve(error || !addresses?.length ? [] : addresses.map(address => address.address))
-      })
-    } catch {
-      resolve([])
-    }
-  })
-}
-
-async function fetchPageHtmlWithCurl(rawUrl: string): Promise<string> {
-  const url = String(rawUrl || '').trim()
-
-  if (!url) {
-    return ''
-  }
-
-  const deadline = Date.now() + TITLE_TIMEOUT_MS
-
-  const result = await fetchWithGuardedRedirects(
-    url,
-    {
-      fetchOnce: (hopUrl, addresses) => curlOneHop(hopUrl, deadline - Date.now(), addresses),
-      resolveHost: resolveHostAddresses
-    },
-    { maxRedirects: PREVIEW_MAX_REDIRECTS }
-  )
-
-  // Refusals (private redirect hop, exhausted redirect budget) and transport
-  // failures both surface as '' — the same miss signal the old curl error
-  // leg produced, so tier 2 and the envelope see nothing new.
-  return result.ok ? result.html : ''
-}
-
-function getLinkTitleSession() {
-  if (linkTitleSession || !app.isReady()) {
-    return linkTitleSession
-  }
-
-  linkTitleSession = session.fromPartition('hermes:link-titles', { cache: false })
-  // One onBeforeRequest registration owns everything this window's requests
-  // must pass: resource blocks AND the B6 SSRF guard (per-request hostname +
-  // DNS verdict, every redirect hop included). A second registration would
-  // replace this one, not stack on it.
-  guardLinkTitleSession(linkTitleSession, { resolveHost: resolveHostAddresses })
-
-  return linkTitleSession
-}
-
-function dequeueRenderTitle() {
-  while (renderTitleInFlight < RENDER_TITLE_MAX_CONCURRENT && renderTitleQueue.length) {
-    const item = renderTitleQueue.shift()
-    renderTitleInFlight += 1
-    runRenderTitleJob(item.url).then(title => {
-      renderTitleInFlight -= 1
-      item.resolve(title)
-      dequeueRenderTitle()
-    })
-  }
-}
-
-function runRenderTitleJob(rawUrl) {
-  return new Promise(resolve => {
-    if (!app.isReady()) {
-      return resolve('')
-    }
-
-    const partitionSession = getLinkTitleSession()
-
-    if (!partitionSession) {
-      return resolve('')
-    }
-
-    let settled = false
-    let window = null
-    let hardTimer = null
-    let graceTimer = null
-
-    const finish = title => {
-      if (settled) {
-        return
-      }
-
-      settled = true
-
-      if (hardTimer) {
-        clearTimeout(hardTimer)
-      }
-
-      if (graceTimer) {
-        clearTimeout(graceTimer)
-      }
-
-      const value = (title || '').replace(/\s+/g, ' ').trim()
-
-      try {
-        if (window && !window.isDestroyed()) {
-          window.destroy()
-        }
-      } catch {
-        // BrowserWindow may already be torn down; ignore.
-      }
-
-      resolve(value)
-    }
-
-    try {
-      window = createLinkTitleWindow(BrowserWindow, partitionSession)
-    } catch {
-      return finish('')
-    }
-
-    const finishWithTitle = () => finish(readLinkTitleWindowTitle(window))
-
-    const scheduleGrace = () => {
-      if (graceTimer) {
-        clearTimeout(graceTimer)
-      }
-
-      graceTimer = setTimeout(finishWithTitle, RENDER_TITLE_GRACE_MS)
-    }
-
-    hardTimer = setTimeout(finishWithTitle, RENDER_TITLE_TIMEOUT_MS)
-
-    window.webContents.setUserAgent(TITLE_USER_AGENT)
-    window.webContents.on('page-title-updated', scheduleGrace)
-    window.webContents.on('did-finish-load', scheduleGrace)
-    window.webContents.on('did-fail-load', (_event, _code, _desc, _validatedURL, isMainFrame) => {
-      if (isMainFrame) {
-        finish('')
-      }
-    })
-
-    window
-      .loadURL(rawUrl, {
-        httpReferrer: 'https://www.google.com/',
-        userAgent: TITLE_USER_AGENT
-      })
-      .catch(() => finish(''))
-  })
-}
-
-function fetchHtmlTitleWithRenderer(rawUrl: string): Promise<string> {
-  return new Promise(resolve => {
-    renderTitleQueue.push({ resolve, url: rawUrl })
-    dequeueRenderTitle()
-  })
-}
-
-// Strips known error/captcha titles (e.g. "GetYourGuide – Error", "Just a
-// moment...") so they don't get cached as the resolved title.
-function usableTitle(value: string): string {
-  return value && !TITLE_ERROR_RE.test(value) ? value : ''
-}
-
-function fetchLinkTitle(rawUrl) {
-  const url = String(rawUrl || '').trim()
-  const key = canonicalTitleCacheKey(url)
-
-  if (!key) {
-    return Promise.resolve('')
-  }
-
-  if (titleCache.has(key)) {
-    return Promise.resolve(titleCache.get(key))
-  }
-
-  if (titleInflight.has(key)) {
-    return titleInflight.get(key)
-  }
-
-  const pending = fetchPageHtmlWithCurl(url)
-    .catch(() => '')
-    .then(value => usableTitle(parseHtmlTitle(value).slice(0, 240)))
-    .then(
-      async value => value || usableTitle(((await fetchHtmlTitleWithRenderer(url).catch(() => '')) || '').slice(0, 240))
-    )
-    .then(clean => {
-      cacheTitle(key, clean)
-      titleInflight.delete(key)
-
-      return clean
-    })
-
-  titleInflight.set(key, pending)
-
-  return pending
-}
-
-// ─── Link previews (D7 click-to-expand unfurl) ───────────────────────────────
-// The policy (SSRF guard, per-hop redirect validation, per-host pacing, cache,
-// field caps) lives in link-preview.ts; this is its I/O and its IPC surface,
-// mirroring how the favicon ladder wires resolveFavicon into main.ts. Tier 1
-// shares fetchPageHtmlWithCurl with the title fetcher — same guarded hops, but
-// the whole byte-budgeted document survives so og/meta tags are readable;
-// tier 2 is the existing hidden-window renderer for pages that only render
-// their title with JS.
-
-const LINK_PREVIEW_CACHE_PATH = path.join(app.getPath('userData'), 'link-preview-cache.json')
-
-const previewPersistence: PreviewPersistence = {
-  read: () => {
-    try {
-      const raw = JSON.parse(fs.readFileSync(LINK_PREVIEW_CACHE_PATH, 'utf8'))
-
-      return raw?.previews && typeof raw.previews === 'object' ? raw.previews : {}
-    } catch {
-      return {}
-    }
-  },
-  write: entries => {
-    try {
-      fs.writeFileSync(LINK_PREVIEW_CACHE_PATH, JSON.stringify({ previews: entries }), 'utf8')
-    } catch {
-      // Cache is an optimization; failing to persist it costs one refetch.
-    }
-  }
-}
-
-const previewStore = new LinkPreviewStore(previewPersistence, { ttlMs: PREVIEW_TTL_MS })
-const previewLimiter = new HostRateLimiter(PREVIEW_HOST_SPACING_MS, PREVIEW_MAX_CONCURRENT)
-
-const previewIo: LinkPreviewIo = {
-  fetchHtml: fetchPageHtmlWithCurl,
-  fetchRenderedTitle: fetchHtmlTitleWithRenderer,
-  resolveHost: resolveHostAddresses
-}
-
-async function fetchLinkPreview(rawUrl: string): Promise<LinkPreviewResult> {
-  return resolveLinkPreview(rawUrl, previewIo, { store: previewStore, limiter: previewLimiter })
-}
 
 // ─── Favicon resolution ──────────────────────────────────────────────────────
 // The ladder itself is electron/favicon.ts; this is its I/O, its cache, and
@@ -16939,9 +16550,10 @@ ipcMain.handle('hermes:setting:defaultProjectDir:pick', async () => {
   return { canceled: false, dir: result.filePaths[0] }
 })
 
-ipcMain.handle('hermes:fetchLinkTitle', (_event, url) => fetchLinkTitle(url))
-
-ipcMain.handle('hermes:fetchLinkPreview', (_event, url) => fetchLinkPreview(url))
+registerLinkPreviewRuntime(
+  ipcMain,
+  createLinkPreviewRuntime({ app, BrowserWindow, session })
+)
 
 ipcMain.handle('hermes:resolveFavicon', (_event, url) => resolveFaviconCached(url))
